Code improvements -Removed junk from InspectorWin class definition -Code, which oriented to ObjectProperties moved from World.cpp to go_properties.cpp -GameObjectLink::crack() implemented
