+++ World/World.cpp
@@ -3,14 +3,7 @@
 #include "headers/Misc.h"
 #include <QLineEdit>
 
-GameObjectProperty::GameObjectProperty(){
-    type = GO_PROPERTY_TYPE_NONE;
-    active = false; //Inactive by default
-}
 
-GameObjectProperty::~GameObjectProperty(){
-
-}
 
 GameObjectLink::GameObjectLink(){
     ptr = nullptr;
@@ -29,35 +22,8 @@ bool GameObjectLink::isEmpty(){
     return false;
 }
 
-void GameObjectProperty::addPropertyInterfaceToInspector(InspectorWin* inspector){
-     //QTextEdit* pos = new QTextEdit;
-    switch(this->type){
-        case GO_PROPERTY_TYPE_TRANSFORM:{ //If it is transform
-            TransformProperty* transfrom = static_cast<TransformProperty*>(this);
-            transfrom->addPropertyInterfaceToInspector(inspector);
-            break;
-        }
-        case GO_PROPERTY_TYPE_LABEL:{ //If it is label
-            LabelProperty* label = static_cast<LabelProperty*>(this);
-            label->addPropertyInterfaceToInspector(inspector);
-            break;
-        }
-    }
-}
-
-void GameObjectProperty::onValueChanged(){
-    switch(this->type){
-        case GO_PROPERTY_TYPE_TRANSFORM:{ //If it is transform
-            TransformProperty* transfrom = static_cast<TransformProperty*>(this);
-            transfrom->onValueChanged();
-            break;
-        }
-        case GO_PROPERTY_TYPE_LABEL:{ //If it is label
-            LabelProperty* label = static_cast<LabelProperty*>(this);
-            label->onValueChanged();
-            break;
-        }
-    }
+void GameObjectLink::crack(){
+    this->world_ptr = nullptr; //It will now pass isEmpty() check
 }
 
 GameObject::GameObject(){
@@ -179,8 +145,9 @@ TransformProperty* GameObject::getTransformProperty(){
 }
 
 void GameObject::addChildObject(GameObjectLink link){
-    link.updLinkPtr();
-    link.ptr->hasParent = true;
+    link.updLinkPtr(); //Calculating object pointer
+    link.ptr->hasParent = true; //Object now has a parent (if it has't before)
+    link.ptr->parent = this->getLinkToThisObject(); //Assigning pointer to new parent
     this->children.push_back(link);
 }
 void GameObject::removeChildObject(GameObjectLink link){
@@ -188,88 +155,9 @@ void GameObject::removeChildObject(GameObjectLink link){
     for(unsigned int i = 0; i < children_am; i++){
         GameObject* ptr = children[i].updLinkPtr();
         if(link.obj_str_id.compare(ptr->str_id) == 0){
-            children[i].world_ptr = nullptr;
-            children[i].ptr = nullptr;
+            children[i].crack(); //Make link broken
         }
     }
-}
-
-TransformProperty::TransformProperty(){
-    type = GO_PROPERTY_TYPE_TRANSFORM; //Type of property is transform
-    active = true; //property is active
-    type_label = "Transform";
-
-    this->transform_mat = getIdentity(); //Result matrix is identity by default
-    this->translation = ZSVECTOR3(0.0f, 0.0f, 0.0f); //Position is zero by default
-    this->scale = ZSVECTOR3(1.0f, 1.0f, 1.0f); //Scale is 1 by default
-    this->rotation = ZSVECTOR3(0.0f, 0.0f, 0.0f); //Rotation is 0 by default
-}
-
-void TransformProperty::addPropertyInterfaceToInspector(InspectorWin* inspector){
-
-    Float3PropertyArea* area_pos = new Float3PropertyArea; //New property area
-    area_pos->setLabel("Position"); //Its label
-    area_pos->vector = &this->translation; //Ptr to our vector
-    area_pos->go_property = static_cast<void*>(this); //Pointer to this to activate matrix recalculaton
-    inspector->addPropertyArea(area_pos);
-
-    Float3PropertyArea* area_scale = new Float3PropertyArea; //New property area
-    area_scale->setLabel("Scale"); //Its label
-    area_scale->vector = &this->scale; //Ptr to our vector
-    area_scale->go_property = static_cast<void*>(this);
-    inspector->addPropertyArea(area_scale);
-
-    Float3PropertyArea* area_rotation = new Float3PropertyArea; //New property area
-    area_rotation->setLabel("Rotation"); //Its label
-    area_rotation->vector = &this->rotation; //Ptr to our vector
-    area_rotation->go_property = static_cast<void*>(this);
-    inspector->addPropertyArea(area_rotation);
-}
-
-void TransformProperty::onValueChanged(){
-    updateMat();
-}
-
-void TransformProperty::updateMat(){
-    //Calculate translation matrix
-    ZSMATRIX4x4 translation_mat = getTranslationMat(this->translation);
-    //Calculate scale matrix
-    ZSMATRIX4x4 scale_mat = getScaleMat(scale);
-    //Calculate rotation matrix
-    ZSMATRIX4x4 rotation_mat = getRotationMat(rotation);
-    //S * R * T
-    this->transform_mat = scale_mat * rotation_mat * translation_mat;
-}
-
-LabelProperty::LabelProperty(){
-    type = GO_PROPERTY_TYPE_LABEL; //its an label
-    active = true;
-}
-
-void LabelProperty::addPropertyInterfaceToInspector(InspectorWin* inspector){
-    StringPropertyArea* area = new StringPropertyArea;
-    area->setLabel("Label");
-    area->value_ptr = &this->label;
-    area->go_property = static_cast<void*>(this);
-    inspector->addPropertyArea(area);
-}
-
-void LabelProperty::onValueChanged(){
-    this->list_item_ptr->setText(0, this->label);
-}
-
-void MeshProperty::addPropertyInterfaceToInspector(InspectorWin* inspector){
-    PickResourceArea* area = new PickResourceArea;
-    area->setLabel("Mesh");
-    area->go_property = static_cast<void*>(this);
-    area->resource_type = RESOURCE_TYPE_MESH; //It should load meshes only
-    inspector->addPropertyArea(area);
-}
-void MeshProperty::update(){
-
-}
-MeshProperty::MeshProperty(){
-
 }
 
 GameObject* World::addObject(GameObject obj){
@@ -347,7 +235,7 @@ void World::saveToFile(QString file){
             unsigned int children_am = static_cast<unsigned int>(object_ptr->children.size());
             for(unsigned int chi_i = 0; chi_i < children_am; chi_i ++){ //iterate over all children
                 GameObjectLink* link_ptr = &object_ptr->children[chi_i]; //Gettin pointer to child
-                if(!link_ptr->isEmpty()){
+                if(!link_ptr->isEmpty()){ //If this link isn't broken (after child removal)
                     world_stream << link_ptr->obj_str_id << " "; //Writing child's string id
                 }
                 }
